refactor(result): 移动类位置
